requirements.txt
discord.py==1.6.0
discord==1.0.1
python-dotenv==0.17.0
